v1,2
In preparation for further code patches. I removed the change in serial
(i.e. OBD) interrupt priority (from 13 back to 1) that was included in
the earlier version because I did not perceive any benefit. I also added
a bit more error checking in the special packet construction routines. I
originally wrote them assuming that I would be the designer of packets
sent to the car and that I wouldn't send erroneous packets. Now, I want
to be sure to protect against malformed packets being sent to the car.